+++ tket2-hseries/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.1](https://github.com/CQCL/tket2/compare/tket2-hseries-v0.14.0...tket2-hseries-v0.14.1) - 2025-06-03
+
+### New Features
+
+- Add V and Vdg to quantum extension. ([#889](https://github.com/CQCL/tket2/pull/889))
+- LLVM codegen for extensions ([#898](https://github.com/CQCL/tket2/pull/898))
+
 ## [0.14.0](https://github.com/CQCL/tket2/compare/tket2-hseries-v0.13.0...tket2-hseries-v0.14.0) - 2025-05-22
 
 ### ⚠ BREAKING CHANGES